+++ launch/nimbus.launch
@@ -1,6 +1,6 @@
 <launch>
     <arg name="config_file" default="$(find nimbus_ros)/config/default.yaml" />
-    <node pkg="nimbus_ros" type="nimbus_ros_node" name="nimbus_ros_node" output="screen" required="true">
+    <node pkg="nimbus_ros" type="nimbus_ros" name="nimbus_ros_node" output="screen" required="true">
         <remap from="pointcloud" to="/nimbus/pointcloud"/>
         <remap from="range_image" to="/nimbus/range_image"/>
         <remap from="intensity_image" to="/nimbus/intensity_image"/>
